% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbSimple.R
\docType{methods}
\name{gibbSimple}
\alias{gibbSimple}
\alias{gibbSimple,matrix,vector,numeric-method}
\title{Execute the gibbs sampler for a simple case}
\usage{
gibbSimple(x, alpha, R)

\S4method{gibbSimple}{matrix,vector,numeric}(x, alpha, R = 1000)
}
\arguments{
\item{x}{adjacency matrix for network}

\item{alpha}{parameters}

\item{R}{number of iterations}
}
\description{
\code{gibbSimple} creates multiple Markov Random Field realizations
for one covariate model. This is done iteratively in \code{Rcpp}
to minimize the computing time.
}
\examples{
N <- 100
x <- matrix(rbinom(N*N,1,0.1),N,N)
alpha <- c(.2,.5)
R <- 1000
output <- gibbSimple(x, alpha, R)

}
