% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbSimplestat.R
\docType{methods}
\name{gibbSimplestat}
\alias{gibbSimplestat}
\alias{gibbSimplestat,matrix,vector,numeric,numeric,vector-method}
\title{Execute the gibbs sampler and save sufficient statistics for a simple case}
\usage{
gibbSimplestat(x, alpha, N, R, weights)

\S4method{gibbSimplestat}{matrix,vector,numeric,numeric,vector}(x, alpha, N, R,
  weights)
}
\arguments{
\item{x}{adjacency matrix for network}

\item{alpha}{vector of alpha parameters}

\item{N}{number of nodes}

\item{R}{number of iterations}

\item{weights}{number of neighbors}
}
\description{
\code{gibbSimplestat} saves sufficient statistics
for a Markov Random Field model. This is done
iteratively in \code{Rcpp} to minimize the computing time.
}
\examples{
N <- 100
x <- matrix(rbinom(N*N,1,0.1),N,N)
alpha <- c(.2,.5)
weights <- apply(x,1,sum)
R <- 100
output <- gibbSimplestat(x, alpha, N, R, weights)

}
