% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsMRF.R
\docType{package}
\name{gibbsMRF}
\alias{gibbsMRF}
\alias{gibbsMRF-package}
\title{gibbsMRF: Gibbs Sampler for an Markov Random Field}
\description{
This function allows for a Gibbs Sampler to run
quickly in Rcpp. For use in the autoGcomputation package.
}
