% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbSimplestat.R
\docType{methods}
\name{gibbSimplestat}
\alias{gibbSimplestat}
\alias{gibbSimplestat,vector,matrix,numeric,numeric,vector,numeric,numeric,vector-method}
\title{Execute the gibbs sampler and save sufficient statistics for a simple case}
\usage{
gibbSimplestat(alpha, x, N, R, weights, burnin, thin, start)


  \S4method{gibbSimplestat}{vector,matrix,numeric,numeric,vector,numeric,numeric,vector}(alpha,
  x, N, R = 500, weights, burnin = 0, thin = 1, start)
}
\arguments{
\item{alpha}{parameters}

\item{x}{adjacency matrix for network}

\item{N}{number of nodes}

\item{R}{number of output iterations}

\item{weights}{number of neighbors}

\item{burnin}{cutoff value for burnin}

\item{thin}{thinning value for chain}

\item{start}{vector of starting values}
}
\description{
\code{gibbSimplestat} saves sufficient statistics
for a Markov Random Field model. This is done
iteratively in \code{Rcpp} to minimize the computing time.
}
\examples{
N <- 100
x <- matrix(rbinom(N*N,1,0.1),N,N)
alpha <- c(.2,.5)
weights <- apply(x,1,sum)
output <- gibbSimplestat(alpha, x, N, 100, weights, 10, 2, rbinom(N,1,runif(1,0,1)))

}
